+++ src/lib/QuestList/QuestMainDrops.svelte
@@ -7,7 +7,7 @@
 <script lang="ts">
     const dispatch = createEventDispatcher();
     import type { Language, QuestBuild2Results, QuestItem } from "$lib/api/api.d";
-    export let id : string;
+    export let id : string; // quest id
     export let build_results : QuestBuild2Results;
     const language : Language = user.region.get();
     const drops : QuestItem[] = questAPI.drops(id, language);
@@ -30,6 +30,13 @@
         // no memory piece
         total_drops = drops;
     }
+    function getMissingString(item_id : string) {
+        const missing = build_results.required[item_id] || 0;
+        if (missing <= 0) {
+            return `-`;
+        }
+        return `${missing}/${build_results.required_clean[item_id] || 0}`;
+    }
 </script>
 
 <div class="inline-flex flex-row gap-1 mr-[1vw]" {...$$restProps}>
@@ -39,12 +46,28 @@
                 <QuestItemButton id={drop.item}
                     priority={build_results.priority_items.includes(drop.item)}
                     disabled={!required_keys.includes(drop.item)}
+                    drop_rate={drop.drop_rate}
                     on:click={() => dispatch("click", { data: drop })}
                 />
-                <strong class:invisible={!required_keys.includes(drop.item)}>
-                    {drop.drop_rate}%
-                </strong>
+                <small
+                    class="required-amount"
+                    class:invisible={!required_keys.includes(drop.item)}
+                >
+                    {getMissingString(drop.item)}
+                </small>
             </div>
         {/if}
     {/each}
 </div>
+
+<style>
+    .required-amount {
+        color: white;
+        font-weight: 800;
+        font-family: "Calibri", Arial, serif;
+        text-shadow: 1px 1px 2px #000000,
+            1px 1px 2px #000000,
+            1px 1px 4px #000000,
+            1px 1px 4px #000000;
+    }
+</style>