+++ src/lib/QuestList/QuestSubDrops.svelte
@@ -12,6 +12,13 @@
     const language : Language = user.region.get();
     const subdrops : QuestItem[] = questAPI.subdrops(id, language);
     const required_keys = Object.keys(build_results.required);
+    function getMissingString(item_id : string) {
+        const missing = build_results.required[item_id] || 0;
+        if (missing <= 0) {
+            return `-`;
+        }
+        return `${missing}/${build_results.required_clean[item_id] || 0}`;
+    }
 </script>
 
 <div class="inline-flex flex-row gap-1 mr-[1vw]" {...$$restProps}>
@@ -20,11 +27,27 @@
             <QuestItemButton id={drop.item}
                 priority={build_results.priority_items.includes(drop.item)}
                 disabled={!required_keys.includes(drop.item)}
+                drop_rate={drop.drop_rate}
                 on:click={() => dispatch("click", { data: drop })}
             />
-            <strong class:invisible={!required_keys.includes(drop.item)}>
-                {drop.drop_rate}%
-            </strong>
+            <small
+                class="required-amount"
+                class:invisible={!required_keys.includes(drop.item)}
+            >
+                {getMissingString(drop.item)}
+            </small>
         </div>
     {/each}
 </div>
+
+<style>
+    .required-amount {
+        color: white;
+        font-weight: 800;
+        font-family: "Calibri", Arial, serif;
+        text-shadow: 1px 1px 2px #000000,
+            1px 1px 2px #000000,
+            1px 1px 4px #000000,
+            1px 1px 4px #000000;
+    }
+</style>